requirements.txt
pyrogram
TgCrypto
python-dotenv
requests
